Moved server JAR to libs directory and modified travis script to run both JARs
