+++ .gitignore
@@ -116,7 +116,12 @@ hs_err_pid*
 
 ### Gradle ###
 .gradle
-build/
+build/classes
+build/generated
+build/reports
+build/test-results
+build/tmp
+build/.DS_Store
 
 # Ignore Gradle GUI config
 gradle-app.setting
@@ -131,6 +136,6 @@ gradle-app.setting
 # gradle/wrapper/gradle-wrapper.properties
 
 ### Gradle Patch ###
-**/build/
+#**/build/
 
 # End of https://www.gitignore.io/api/java,gradle,intellij+all
