+++ .travis.yml
@@ -3,5 +3,6 @@ language: java
 jdk: openjdk11
 
 script:
-  - java -jar build/libs/httpApplication-1.0-SNAPSHOT.jar &
+  - java -cp httpApplication-1.0-SNAPSHOT.jar:build/libs/*:. App & MyPID=$!
   - "./acceptance-tests.sh"
+  - kill $MyPID
