+++ build.gradle
@@ -12,7 +12,7 @@ repositories {
 
 dependencies {
     testCompile group: 'junit', name: 'junit', version: '4.12'
-    implementation files('dependencies/httpServer-1.0-SNAPSHOT.jar')
+    implementation files('build/libs/httpServer-1.0-SNAPSHOT.jar')
 //    implementation files('../http-server-make-server-dependent/build/libs/httpServer-1.0-SNAPSHOT.jar')
 }
